+++ pkglogger.test.js
@@ -2,20 +2,21 @@ var pkglogger = require('./pkglogger');
 
 var log = pkglogger(module);
 
+var msg = 1;
+
 function logMessages() {
-    log.trace('This is a trace message.');
-    log.debug('This is a debug message.');
-    log.info('This is an info message.');
-    log.warn('This is a warn message.');
-    log.error('This is an error message.');
-    log.fatal('This is a fatal message.');
+    log.trace('This is trace message {0}.', msg++);
+    log.debug('This is debug message {0}.', msg++);
+    log.info('This is info message {0}.', msg++);
+    log.warn('This is warn message {0}.', msg++);
+    log.error('This is error message {0}.', msg++);
+    log.fatal('This is fatal message {0}.', msg++);
 }
 
 logMessages();
-log.setLevel(log.TRACE);
+log.level(log.TRACE);
 logMessages();
-log.setLevel(log.OFF);
-log.useStderr(true);
+log.level(log.OFF).stderr(true);
 logMessages();
-log.setLevel(log.WARN);
+log.level(log.WARN);
 logMessages();